Update Terragrunt to latest version & Go Mod Tidy
As mentioned in https://github.com/transcend-io/terragrunt-atlantis-config/issues/102
there is an issue in Terragrunt version <= 0.23.31 .
As the project uses this old version it seems like a good pattern to update
the Terragrunt version to latest one.
The tests in the project are PASSing and also a project that has multiple
environments that was failing for the 0.13.0 that uses the older terragrunt
version i believe there should be no side-effects introduced by this change.